pyproject.toml
[project]
name = "video-analyzer"
version = "0.1.0"
description = "YouTube video analyzer using Gemini/Groq API for transcription, vision analysis, and hook detection"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "groq>=0.33.0",
    "yt-dlp>=2024.1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "typer>=0.9.0",
    "sqlmodel>=0.0.14",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "pillow>=10.2.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",
    "ruff>=0.14.13",
]

[project.scripts]
video-analyzer = "video_analyzer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video_analyzer"]
